refactor: reorganize assets
