+++ src/assets/js/lib.d.ts
